Rename files and format with black formatter
