+++ activity.py
@@ -2,7 +2,8 @@ from gettext import gettext as _
 
 import sys
 import gi
-gi.require_version('Gtk', '3.0')
+
+gi.require_version("Gtk", "3.0")
 from gi.repository import Gtk
 import pygame
 
@@ -13,15 +14,14 @@ from sugar3.graphics.toolbutton import ToolButton
 from sugar3.activity.widgets import StopButton
 
 
-
 import sugargame.canvas
 
-import TestGame
+import game
 
 
-class TestActivity(sugar3.activity.activity.Activity):
+class BallAndBrickActivity(sugar3.activity.activity.Activity):
     def __init__(self, handle):
-        super(TestActivity, self).__init__(handle)
+        super(BallAndBrickActivity, self).__init__(handle)
 
         self.paused = False
 
@@ -29,19 +29,18 @@ class TestActivity(sugar3.activity.activity.Activity):
         self.build_toolbar()
 
         # Create the game instance.
-        self.game = TestGame.TestGame()
+        self.game = game.BallAndBrick()
 
         # Build the Pygame canvas.
 
-        self._pygamecanvas = sugargame.canvas.PygameCanvas(self,
-            main=self.game.run,
-            modules=[pygame.display])
+        self._pygamecanvas = sugargame.canvas.PygameCanvas(
+            self, main=self.game.run, modules=[pygame.display]
+        )
 
         # Note that set_canvas implicitly calls read_file when
         # resuming from the Journal.
         self.set_canvas(self._pygamecanvas)
         self._pygamecanvas.grab_focus()
-
 
     def build_toolbar(self):
         toolbar_box = ToolbarBox()
@@ -51,7 +50,6 @@ class TestActivity(sugar3.activity.activity.Activity):
         activity_button = ActivityToolbarButton(self)
         toolbar_box.toolbar.insert(activity_button, -1)
         activity_button.show()
-
 
         # Blank space (separator) and Stop button at the end:
 
@@ -65,12 +63,11 @@ class TestActivity(sugar3.activity.activity.Activity):
         toolbar_box.toolbar.insert(stop_button, -1)
         stop_button.show()
 
-
     def read_file(self, file_path):
         self.game.read_file(file_path)
 
     def write_file(self, file_path):
         self.game.write_file(file_path)
-    
+
     def get_preview(self):
         return self._pygamecanvas.get_preview()
